BAH-4686 | Address PR review comments
