+++ stories/mockData.js
@@ -1,16 +1,9 @@
-/* global componentStore */
-import { NumericBox } from 'src/components/NumericBox.jsx';
-import { BooleanControl } from 'src/components/BooleanControl.jsx';
-import { Button } from 'src/components/Button.jsx';
-import { CodedControl } from 'src/components/CodedControl.jsx';
-import { AutoComplete } from 'src/components/AutoComplete.jsx';
-import { TextBox } from 'src/components/TextBox.jsx';
-
 export const SYSTOLIC_UUID = 'c36e9c8b-3f10-11e4-adec-0800271c1b75';
 export const DIASTOLIC_UUID = 'c379aa1d-3f10-11e4-adec-0800271c1b75';
-export const PULSE_DATA_UUID = 'c36af094-3f10-11e4-adec-0800271c1b75';
-export const PULSE_UUID = 'c36bc411-3f10-11e4-adec-0800271c1b75';
-export const PULSE_ABNORMAL_UUID = 'c36c7c98-3f10-11e4-adec-0800271c1b75';
+
+const PULSE_DATA_UUID = 'c36af094-3f10-11e4-adec-0800271c1b75';
+const PULSE_UUID = 'c36bc411-3f10-11e4-adec-0800271c1b75';
+const PULSE_ABNORMAL_UUID = 'c36c7c98-3f10-11e4-adec-0800271c1b75';
 
 export const pulseDataMetadata = {
   type: 'ObsGroupControl',
@@ -53,12 +46,3 @@ export const pulseDataMetadata = {
     },
   ],
 };
-
-export function registerCoreComponents() {
-  componentStore.registerComponent('numeric', NumericBox);
-  componentStore.registerComponent('boolean', BooleanControl);
-  componentStore.registerComponent('button', Button);
-  componentStore.registerComponent('Coded', CodedControl);
-  componentStore.registerComponent('autoComplete', AutoComplete);
-  componentStore.registerComponent('text', TextBox);
-}
